FInd element in Sorted and Rotated
Used binary search but made some changes on which side to look for and which side not to look for! if its sorted check in that and keep dping that we wont check in the unsorted one we will just assume its in that subarray and we find that sucker in the end.