search in 2d Par2
basically did an o(n + m), else would have been the O(nlog(m)) to do binary search in every row, and iterate through all the rows. Now, we do o(n + m), starting from top right, as top right is the middle element of a sorted line. If we found the element, return true, or if its less, got to left, or go to right. DOing so we go through diagonal size of the array, which is basically o(n+m).